move com->itc
