+++ client/src/connection/itc/script.ts
